Changes to file structure. Moving testingfiles in to the correct folders and giving them correct naming

+++ src/js/api/auth/sign-out.js
@@ -1,6 +1,8 @@
 export function signOutBtn() {
   const signOutBtn = document.querySelector('#signOutBtn');
-  signOutBtn.addEventListener('click', () => {
-    localStorage.clear('Token', 'Profile');
-  });
+  if (signOutBtn) {
+    signOutBtn.addEventListener('click', () => {
+      localStorage.clear('Token', 'Profile');
+    });
+  }
 }
